docker.run {epimod}	R Documentation

Run docker container

Description

This is an internal function executing a docker container. Not to be used by users.

Usage

docker.run(params = NULL, changeUID = TRUE, debug = FALSE)

Arguments

params,
a character string containing all parameters needed to run the docker container.

Value

0 if success, 1 if parameters are missing, 2 if dockerid file is present, 3 if docker execution fails.

Author(s)

Beccuti Marco, Castagno Paolo, Pernice Simone

Examples

## Not run: 
    #running runDocker
     docker.run(params=NULL, changeUID=TRUE, debug=FALSE)

## End(Not run)

[Package epimod version 1.0.0 Index]