downloadContainers {epimod}	R Documentation

Download for the first time all containers embedded in the workflows

Description

This is a function that prepares the docker environment to be used for the first time the epimod is installed.

Usage

downloadContainers(containers.file = NULL, tag = NULL)

Arguments

containers.file,
a character string with the name of the file which indicate which are the initial set of containers to be downloaded. If NULL then the set is given by a file called "containersNames.txt" located in the folder inst/Containers of epimod package.

Author(s)

Beccuti Marco, Castagno Paolo, Pernice Simone

Examples

## Not run: 
    #running runDocker
     downloadContainers()

## End(Not run)

[Package epimod version 1.0.0 Index]