experiment.cmd {epimod}	R Documentation

Generate the command template to configure and run a simulation

Description

Function generating the command line used to run the solver with the appropriate configuration

Usage

experiment.cmd(
  solver_fname,
  solver_type = "LSODA",
  taueps = 0.01,
  seed,
  timeout
)

Arguments

solver_fname,
the solver executable file

solver_type,
the type of solver (LSODA, HLSODA, ..)

taueps,
controls the time step for tau-leap algorithm

timeout,
available time to run n_run simulations (See TIMEOUT(1) to check the syntax)

event_times,
times at which the marking is updated through the event_function

event_function,
specifies the rule to update the marking

out_fname,
output filename prefix

Value

a string

Author(s)

Paolo Castagno, Simone Pernice

Examples

## Not run: 
experiment.cmd(id=1, solver_fname="Solver.solver", s_time=365, f_time=730, timeout="1d", out_fname="simulation")

## End(Not run)

[Package epimod version 1.0.0 Index]