experiment.configurations {epimod}	R Documentation

Generates a set of simulation's configurations

Description

This is an internal function generating a set of configurations. The generation process is driven by the functions listed in param_list. Such functions either are native R functions or user defined ones. In the latter case, the functions have to be defined in a single file and passed through the argument parm_fname

Usage

experiment.configurations(
  n_config,
  parm_fname = NULL,
  parm_list = NULL,
  out_dir,
  out_fname,
  ini_vector = NULL,
  ini_vector_mod = FALSE,
  extend = FALSE,
  config = list()
)

Arguments

n_config,
number of configuratons to generate

parm_fname,
file with the definition of user defined functions

parm_list,
file listing the name of the functions, the parameters and the name under which the parameters have to be saved

out_dir,
output directory specified by the user

out_fname,
prefix used to name output files

extend,
unused parameter. Placeholder for future development

Author(s)

Marco Beccuti, Paolo Castagno, Simone Pernice

Examples

## Not run: 
experiment.configurations(nconfig = 10,
                          param_list = list.csv,
                          out_dir="/path/to/output/directory",
                          out_fname="example")

## End(Not run)

[Package epimod version 1.0.0 Index]