experiment.env_setup {epimod}	R Documentation

Sets up the environment (directory and files) required to run a simulartion

Description

This is an internal function creating the required structure of files and directiry to run a simulation.

Usage

experiment.env_setup(id = NULL, files, dest_dir, config = NULL)

Arguments

id,
a numeric identifier used to format the output's file name (optional, if not provided copy files in dest_dir. Otherwise, create a sub-directory id)

confg,
configuration provided by the function experiment.configurations

Author(s)

Beccuti Marco, Castagno Paolo, Pernice Simone

Examples

## Not run: 
experiment.env_cleanup(id=1, run_dir="/run/directory", out_fname="simulation", out_dir="/out/directory",)

## End(Not run)

[Package epimod version 1.0.0 Index]