model.generation {epimod}	R Documentation

Run model generation

Description

Having constructed the model exploiting the graphical editor, namely GreatSPN, the automatic generation of both the stochastic (the Continuous Time Markov Chain) and deterministic (ODEs) processes underlying the model is implemented by the model_generation() function.

Usage

model.generation(
  out_fname = NULL,
  net_fname,
  transitions_fname = NULL,
  LP = F,
  volume = getwd(),
  debug = FALSE
)

Arguments

out_fname
Prefix to the output file name.

net_fname
.PNPRO file storing the Petri Net (and all its generalizations) model. In case there are multiple nets defined within the PNPRO file, the first one in the list is the will be automatically selected.

transitions_fname
C++ file defining the functions managing the behaviour of general transitions, mandatory if Extended versions of Petri Nets (i.e., ESPN or ESSN) are used.

LP
Enabling the FBA solution encoded in the general transitions. (default is FALSE)

volume
The folder to mount within the Docker image providing all the necessary files.

debug
If TRUE enables logging activity.

Details

GreatSPN GUI, the graphical editor for drawing Petri Nets formalism, is available online: http://www.di.unito.it/~amparore/mc4cslta/editor.html

Value

model.generation returns the binary file SIR.solver in which the underlying processes (both deterministic and stochastic) of the Petri Net model and the library used for their simulation are packaged.

Author(s)

Beccuti Marco, Castagno Paolo, Pernice Simone, Baccega Daniele

Examples

## Not run: 
local_dir <- "/some/path/to/the/directory/hosting/the/input/files/"
model.generation(out_fname = "Solver",
                 net_fname = paste0(local_dir, "Configuration/Pertussis"),
                 transitions_fname = "transitions.cpp")

## End(Not run)

[Package epimod version 1.0.0 Index]